워드 클라우드 API 상세 수치 설정하기 (React JS and Firebase Web App Project #12)

+++ word_cloud.py
@@ -51,11 +51,11 @@ def make_cloud_image(tags, file_name):
 
 
 def process_from_text(text, max_count, min_length, words, file_name):
-    tags = get_tags(text, max_count, min_length)
+    tags = get_tags(text, int(max_count), int(min_length))
     # 단어 가중치를 적용합니다.
     for n, c in words.items():
         if n in tags:
-            tags[n] = tags[n] * int(words[n])
+            tags[n] = tags[n] * float(words[n])
     make_cloud_image(tags, file_name)
